size_t network_send_bytes

+++ network.h
@@ -110,7 +110,7 @@ extern int network_send_line(network_handle nh, const char *line,
 				 */
 
 extern int network_send_bytes(network_handle nh,
-			      const char *buffer, int buflen,
+			      const char *buffer, size_t buflen,
 			      int flush_ok);
 				/* The first BUFLEN bytes in the given BUFFER
 				 * should be queued for output on the specified
